refactor: BookDocument를 infra 패키지로 이동
