+++ src/main/java/com/bookspot/book/infra/search/result/BookPageResult.java
@@ -1,6 +1,6 @@
 package com.bookspot.book.infra.search.result;
 
-import com.bookspot.book.infra.search.BookDocument;
+import com.bookspot.book.infra.BookDocument;
 import org.springframework.data.domain.Page;
 
 public record BookPageResult(Page<BookDocument> books, String lastScore, Integer lastLoanCount, Long lastBookId) {
